@@ -1,11 +1,11 @@
 name: flutter_slidable
 description: A Flutter implementation of slidable list item with directional slide actions that can be dismissed.
-version: 3.1.2
+version: 4.0.0
 homepage: https://github.com/letsar/flutter_slidable
 
 environment:
-  sdk: '>=2.17.0 <4.0.0'
-  flutter: ">=3.7.0"
+  sdk: '>=3.6.0 <4.0.0'
+  flutter: ">=3.27.0"
 
 dependencies:
   flutter:
